+++ CHANGELOG.rst
@@ -10,6 +10,9 @@ Changelog
 - Add base log path option: `-b` `--log-base`.
 - Log postfix option: `-p` `--log-postfix`.
 - Auto alias generation: `-a` `--alias-level`.
+- Add support for default /var/log/messages datetime format files with
+  `-ml [FILE [FILE]]`
+- Add support for timestamped log files with `-tl [FILE [FILE]]`
 
 **Bugfixes:**